requirements.txt
PyYAML==3.13
requests==2.21
flask
flask-restful